Support extension services
Several changes to better support extension services:

* Use Guice to publish the list of known node roles so
  an extension can add to it.
* Use the Guice list for the SQL "servers" system table so
  it will publish instances of the extension service.
* Split out the index task client into a base class to
  allow the client to be used for an extension service.
* Add a PrivateApi annotation to mark REST services primarly
  used internally to the server.
* Misc. code cleanup.
