+++ indexing-service/src/main/java/org/apache/druid/indexing/common/SingleFileTaskReportFileWriter.java
@@ -53,7 +53,6 @@ public class SingleFileTaskReportFileWriter implements TaskReportFileWriter
     }
   }
 
-
   @Override
   public void setObjectMapper(ObjectMapper objectMapper)
   {
