+++ sql/src/test/java/org/apache/druid/sql/calcite/util/CalciteTests.java
@@ -53,6 +53,7 @@ import org.apache.druid.discovery.DruidLeaderClient;
 import org.apache.druid.discovery.DruidNodeDiscovery;
 import org.apache.druid.discovery.DruidNodeDiscoveryProvider;
 import org.apache.druid.discovery.NodeRole;
+import org.apache.druid.discovery.NodeRoles;
 import org.apache.druid.guice.ExpressionModule;
 import org.apache.druid.guice.annotations.Json;
 import org.apache.druid.java.util.common.DateTimes;
@@ -1155,6 +1156,7 @@ public class CalciteTests
         druidLeaderClient,
         overlordLeaderClient,
         provider,
+        NodeRoles.knownRoles(),
         getJsonMapper()
     );
   }
